[#141] Add taquito lib and implement sample contract calls
Problem: We have to implement some Typescript code to call the BaseDAO
contract, hopefully after finding a way to wrap its parameter into some
nice Typescript types.

Solution: It was found that there is a Typescript library called
`taquito` that already provides the capability of calling Tezos smart
contracts over RPC. So here we start a Typescript library including
`taquito` library in the dependencies. We have also included a sample
code that calls a contract here as well.