[#141] Add typescript code generator
Problem: It would be nice to have the types that wraps the parameter of
the BaseDAO contract to be generated from the contract parameter itself.
If we put these type in the api interface, and its implementation, we
can have some type level guarantee that the api and the contract is not
out of sync.

Solution: Use the parameter notes generated by Morley to generate
Typescript types that wraps the entrypoints types.