[#151] Store more info for RegistryDAO
Problem: We need to add a `name` and `field` field to each proposal. We
also need to be able to figure out if some voter voted for or against a
proposal.

Solution: For the relevant Proposal types, add a name and description of
`string`/`MText` types. For the relevant voters, change the `Natural` so
that it's a right-combed `(Natural, VoteType`), which allows us to
determine more about the vote. Create a name and description for each
existing test. Add a new test that checks whether we can only vote iff
the name and description match.